Time: 22 ms (42.88%) | Memory: 45.7 MB (9.17%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.java
@@ -18,7 +18,7 @@ class Solution {
         int currHrs = 0;
 
         for (int pile : piles) {
-            currHrs += Math.ceil(( (double) pile / k));
+            currHrs += (int) Math.ceil(( (double) pile / k));
             if (currHrs > reqHrs) return false;
         }
